+++ R/order_by.R
@@ -21,15 +21,17 @@
 #'
 #' @examples
 #'
-#' my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
-#' d <- dbi_copy_to(my_db, 'd',
-#'                 data.frame(AUC = 0.6, R2 = 0.2))
-#' eqn <- orderby(d, rev_cols = "AUC", limit=4)
-#' cat(format(eqn))
-#' sql <- to_sql(eqn, my_db)
-#' cat(sql)
-#' DBI::dbGetQuery(my_db, sql)
-#' DBI::dbDisconnect(my_db)
+#' if (requireNamespace("RSQLite", quietly = TRUE)) {
+#'   my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+#'   d <- dbi_copy_to(my_db, 'd',
+#'                    data.frame(AUC = 0.6, R2 = 0.2))
+#'   eqn <- orderby(d, rev_cols = "AUC", limit=4)
+#'   cat(format(eqn))
+#'   sql <- to_sql(eqn, my_db)
+#'   cat(sql)
+#'   print(DBI::dbGetQuery(my_db, sql))
+#'   DBI::dbDisconnect(my_db)
+#' }
 #'
 #' @export
 #'
@@ -194,10 +196,3 @@ to_sql.relop_orderby <- function (x,
   c(subsql_list[-length(subsql_list)], q)
 }
 
-#' @export
-#'
-dim.relop_orderby <- function(x) {
-  ncol <- length(column_names(x))
-  nrow <- nrow(x$source[[1]])
-  c(nrow, ncol)
-}