+++ R/rename.R
@@ -7,15 +7,17 @@
 #'
 #' @examples
 #'
-#' my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
-#' d <- dbi_copy_to(my_db, 'd',
-#'                 data.frame(AUC = 0.6, R2 = 0.2, z = 3))
-#' eqn <- rename_columns(d, c('AUC2' := 'AUC', 'R' := 'R2'))
-#' cat(format(eqn))
-#' sql <- to_sql(eqn, my_db)
-#' cat(sql)
-#' DBI::dbGetQuery(my_db, sql)
-#' DBI::dbDisconnect(my_db)
+#' if (requireNamespace("RSQLite", quietly = TRUE)) {
+#'   my_db <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
+#'   d <- dbi_copy_to(my_db, 'd',
+#'                    data.frame(AUC = 0.6, R2 = 0.2, z = 3))
+#'   eqn <- rename_columns(d, c('AUC2' := 'AUC', 'R' := 'R2'))
+#'   cat(format(eqn))
+#'   sql <- to_sql(eqn, my_db)
+#'   cat(sql)
+#'   print(DBI::dbGetQuery(my_db, sql))
+#'   DBI::dbDisconnect(my_db)
+#' }
 #'
 #' @export
 #'
@@ -172,11 +174,5 @@ to_sql.relop_rename_columns <- function (x,
   c(subsql_list[-length(subsql_list)], q)
 }
 
-#' @export
-#'
-dim.relop_rename_columns <- function(x) {
-  ncol <- length(column_names(x))
-  nrow <- nrow(x$source[[1]])
-  c(nrow, ncol)
-}
+