+++ NEWS.md
@@ -1,4 +1,14 @@
 
+# rquery 0.4.1 2018/03/17
+
+ * Fix re-run/re-create situations.
+ * Pipe database connection.
+ * Better quantile calc (assume window functions).
+ * Improve dependency decls.
+ * Further limit direct dependencies.
+ * Work with more DB drivers.
+ * SQL node for sets of columns.
+
 # rquery 0.4.0 2018/03/10
 
  * Move to wrapr 1.2.0